Fix user update function
If user didnt send the email in the update json server identified an error.

User could also change password without sending the old password. Now, the logic is inverted in Yup (if user send the password, oldPassword and confirmPassword are required. If he send the other two, nothing changes on the database).